requirements.txt
gin-config==0.5.0
numpy==1.23.5
Pillow==9.3.0
tensorboard==2.11.0
timm==0.6.12
torch==1.13.0
torchattacks==3.3.0
torchmetrics==0.11.0
torchvision==0.14.0
tqdm==4.64.1
scipy==1.9.3
